Type counts and durations as constrained integers
catalog.types.ck already said int(min=0) for durationMs and the counts; playout,
playlists and plugins said bare number for the same values. One spelling now,
which also gets .int().min(0) onto the generated Zod.

Checked the sources before tightening: radio.liq already floors remainingMs with
int_of_float, startedAt is Date.now(), and durationMs comes from Spotify's
duration_ms and MusicBrainz's length, both integer milliseconds.

bpm stays number - a tempo a source measured rather than declared is fractional,
and the contract already carries that comment.

+++ apps/api/src/modules/plugins/types/plugins.types.ts
@@ -119,7 +119,7 @@ export type PluginLogEntry = z.infer<typeof PluginLogEntry>;
  * generated from [PluginLogQuery](file://./../../../../data/contracts/plugins/plugins.types.ck#L59)
  */
 export const PluginLogQuery = z.strictObject({
-    limit: z.coerce.number().min(1).max(2000).optional(),
+    limit: z.coerce.number().int().min(1).max(2000).optional(),
     level: PluginLogLevel.optional(),
 });
 export type PluginLogQuery = z.infer<typeof PluginLogQuery>;
